Create GNOME palette.md
This file contains common colors used in dark adwaita. Later on I may create the theme for default adwaita theme. But it will be later and I am not sure, because GNOME 42 may roll out in Fedora 36 soon and it has a different color theme and also GTK Telegram clients are already in development.